+++ sympy/functions/elementary/tests/test_trigonometric.py
@@ -1,4 +1,4 @@
-from sympy.calculus.util import AccumBounds
+from sympy.calculus.accumulationbounds import AccumBounds
 from sympy.core.add import Add
 from sympy.core.function import (Lambda, diff)
 from sympy.core.mul import Mul
