+++ sympy/series/limitseq.py
@@ -1,5 +1,6 @@
 """Limits of sequences"""
 
+from sympy.calculus.accumulationbounds import AccumulationBounds
 from sympy.core.add import Add
 from sympy.core.function import PoleError
 from sympy.core.power import Pow
@@ -203,7 +204,6 @@ def limit_seq(expr, n=None, trials=5):
     """
 
     from sympy.concrete.summations import Sum
-    from sympy.calculus.util import AccumulationBounds
     if n is None:
         free = expr.free_symbols
         if len(free) == 1:
