+++ sympy/sets/handlers/mul.py
@@ -2,6 +2,7 @@ from sympy.core import Basic, Expr
 from sympy.core.numbers import oo
 from sympy.core.symbol import symbols
 from sympy.multipledispatch import dispatch
+from sympy.sets.setexpr import set_mul
 from sympy.sets.sets import Interval, Set
 
 _x, _y = symbols("x y")
@@ -60,7 +61,6 @@ def _set_div(x, y): # noqa:F811
     Divisions in interval arithmetic
     https://en.wikipedia.org/wiki/Interval_arithmetic
     """
-    from sympy.sets.setexpr import set_mul
     if (y.start*y.end).is_negative:
         return Interval(-oo, oo)
     if y.start == 0:
