+++ sympy/sets/powerset.py
@@ -4,7 +4,7 @@ from sympy.core.logic import fuzzy_bool
 from sympy.core.singleton import S
 from sympy.core.sympify import _sympify
 
-from .sets import Set
+from .sets import Set, FiniteSet
 
 
 class PowerSet(Set):
@@ -35,8 +35,7 @@ class PowerSet(Set):
     Examples
     ========
 
-    >>> from sympy.sets.powerset import PowerSet
-    >>> from sympy import S, FiniteSet
+    >>> from sympy import PowerSet, S, FiniteSet
 
     A power set of a finite set:
 
@@ -103,7 +102,6 @@ class PowerSet(Set):
         return 2 ** len(self.arg)
 
     def __iter__(self):
-        from .sets import FiniteSet
         found = [S.EmptySet]
         yield S.EmptySet
